fix(infra): set APP_MODE before the deploy implementation's import graph loads
ESM hoists imports, so the engine-config seam evaluated the shared appConfig
(defaulting to development) before runDeploy could set the mode. deploy.ts is
now a thin loader with no app imports: parse --mode, set APP_MODE, then
dynamically import the implementation (deploy-run.ts).
